# back/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API documentation for Easy DEVS.",
		"title": "Easy DEVS API",
		"contact": {},
		"version": "1.0"
	},
	"host": "localhost:3000",
	"basePath": "/",
	"paths": {
...
		"/auth/login": {
			"post": {
				"description": "Logs in a user with the provided identity (email or username) and password.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["Auth"],
				"summary": "Log in a user",
				"parameters": [
					{
						"description": "Login Request",
						"name": "input",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.LoginRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/response.LoginResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"description": "Logs out a user by invalidating the refresh token.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["Auth"],
				"summary": "Log out a user",
				"parameters": [
					{
						"description": "Logout Request",
						"name": "input",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.LogoutRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"description": "Returns the authenticated user's information based on the access token.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["Auth"],
				"summary": "Get current user",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/response.UserResponse"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"description": "Refreshes the access token using a valid refresh token.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["Auth"],
				"summary": "Refresh access token",
				"parameters": [
					{
						"description": "Refresh Request",
						"name": "input",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.RefreshRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/response.RefreshResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"description": "Registers a new user with the provided username, email, and password.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["Auth"],
				"summary": "Register a new user",
				"parameters": [
					{
						"description": "Register Request",
						"name": "input",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.RegisterRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/response.RegisterResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"409": {
						"description": "Conflict",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"request.LoginRequest": {
			"type": "object",
			"properties": {
				"identity": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			}
		},
		"request.LogoutRequest": {
			"type": "object",
			"properties": {
				"refreshToken": {
					"type": "string"
				}
			}
		},
...
		"request.RefreshRequest": {
			"type": "object",
			"properties": {
				"refreshToken": {
					"type": "string"
				}
			}
		},
		"request.RegisterRequest": {
			"type": "object",
			"required": ["email", "password", "username"],
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string",
					"maxLength": 50,
					"minLength": 6
				},
				"username": {
					"type": "string",
					"maxLength": 50,
					"minLength": 3
				}
			}
		},
...
		"response.LoginResponse": {
			"type": "object",
			"properties": {
				"accessToken": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"refreshToken": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		},
...
		"response.RefreshResponse": {
			"type": "object",
			"properties": {
				"accessToken": {
					"type": "string"
				}
			}
		},
		"response.RegisterResponse": {
			"type": "object",
			"properties": {
				"accessToken": {
					"type": "string"
				},
				"refreshToken": {
					"type": "string"
				},
				"user": {
					"$ref": "#/definitions/response.UserResponse"
				}
			}
		},
		"response.UserResponse": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		}
	}
}